% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cor.R
\name{calc_cor_target}
\alias{calc_cor_target}
\alias{calc_corr_target}
\title{Calculate correlations against a target variable}
\usage{
calc_corr_target(
  df,
  target,
  use = "pairwise.complete.obs",
  fct_reorder = FALSE,
  fct_rev = FALSE
)
}
\arguments{
\item{df}{A tibble}

\item{target}{Name of continuous target variable}

\item{use}{A string indicating missing valuable treatment for correlation analysis}

\item{fct_reorder}{Re-order factors?}

\item{fct_rev}{Reverse factors?}
}
\description{
Calculate correltions for all numeric variables against a target variable.
}
\examples{
get_cor(mtcars, mpg)

}
