% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_corr_group.R
\name{calc_corr_group}
\alias{calc_corr_group}
\title{Calculate correlations by group}
\usage{
calc_corr_group(df, cont_var_1, cont_var_2, ...)
}
\arguments{
\item{df}{A tibble}

\item{cont_var_1}{A numeric vector (continuous variable)}

\item{cont_var_2}{A numeric vector (continuous variable)}

\item{...}{One or more categorical variables}
}
\description{
For two continuous variables, get correlations by leves of a categorical variable
}
