% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_count.R
\name{calc_count}
\alias{calc_count}
\title{Get frequency and percentage counts for categorical variables}
\usage{
calc_count(df, ..., col = n, show_percent = TRUE)
}
\arguments{
\item{df}{A tibble.}

\item{col}{A fixed value that is quoted, representing number of cases.}

\item{show_percent}{Round and format percentages with \%.}
}
\description{
Calculate counts & percentages for selection of categorical variables
}
\examples{
mtcars \%>\% calc_count(cyl)
mtcars \%>\% calc_count(cyl, am)

}
