% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_count.R
\name{calc_count_all}
\alias{calc_count_all}
\title{Calculate All Frequency & Percentage Counts}
\usage{
calc_count_all(df, col = n, show_percent = TRUE, max_categories = 5)
}
\arguments{
\item{df}{A tibble.}

\item{col}{A fixed value that is quoted, representing number of cases.}

\item{show_percent}{Round and format percentages with \%.}

\item{max_categories}{A numeric scalar. Treat numeric variable as categorical for max_categories.}
}
\description{
Calculate frequency and percentage caounts for all categorical variables.
}
\examples{
mtcars \%>\% get_counts_all(max_categories = 3)

}
