% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crosscount.R
\name{calc_crosscount}
\alias{calc_crosscount}
\title{Crosstablulate Counts & Percentages}
\usage{
calc_crosscount(df, grp_var_1, grp_var_2)
}
\arguments{
\item{df}{A tibble}

\item{grp_var_1}{A categorical variable}

\item{grp_var_2}{A categorical variable}
}
\description{
Breakdown counts and percentages of one categorical variable by another.
}
\examples{
calc_crosscount(mtcars, am, cyl)

}
