% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crosscount.R
\name{calc_crosscount_all}
\alias{calc_crosscount_all}
\title{Caclculate Counts and Percentages for All Categorical Variables Against Another}
\usage{
calc_crosscount_all(df, target_var, max_categories = 5)
}
\arguments{
\item{df}{A tibble}

\item{target_var}{A categorical variable to target in analysis.}

\item{max_categories}{A numeric scalar. Treat numeric variable as categorical for max_categories.}
}
\description{
For a target categorical variable, get counts and percentages for levels of all categorical variables.
}
\examples{
calc_crosscount_all(mtcars, cyl)

}
