% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_group_mean.R
\name{calc_group_mean}
\alias{calc_group_mean}
\title{Calculate Grouped Means}
\usage{
calc_group_mean(df, cont_var, ...)
}
\arguments{
\item{df}{A tibble}

\item{cont_var}{A continuous variable}
}
\description{
Calculate means by one or more grouping variables.
}
\examples{
calc_group_mean(mtcars, mpg, cyl)

}
