% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sig_corr_all.R
\name{calc_sig_corr_all}
\alias{calc_sig_corr_all}
\title{Calculate all significant correlations}
\usage{
calc_sig_corr_all(df)
}
\arguments{
\item{df}{A tibble}
}
\description{
Calculate all significant correlations in a dataset and store in a tibble.
}
\examples{
calc_sig_corr_all(mtcars)

}
