% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_val.R
\name{format_p_val}
\alias{format_p_val}
\title{Format p-values}
\usage{
format_p_val(p_val_column)
}
\arguments{
\item{p_val_column}{A numeric scalar or vector of p-values}
}
\description{
Format p-values to 5 decimal places
}
\examples{
format_p_val(x)

}
