% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Plot a Boxplot}
\usage{
plot_box(df, cont_var, grp_var = NULL, identifier, interactive = FALSE)
}
\arguments{
\item{df}{A tibble.}

\item{cont_var}{A continuous variable.}

\item{grp_var}{A categorical variable to group by}

\item{identifier}{A variable to identify individual points if interactive = TRUE}

\item{interactive}{If true, plots are passed to plotly and display tooltips.}
}
\description{
Plot boxplots for a continuous variable across zero or one categorical variable.
}
\examples{

plot_box(mtcars, cont_var = mpg, grp_var = cyl, identifier = am)

}
