% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box_all}
\alias{plot_box_all}
\title{Plot All Boxplots}
\usage{
plot_box_all(
  df,
  grp_var = NULL,
  identifier,
  interactive = FALSE,
  n_rows = 4,
  margins = c(0.059, 0.02, 0.02, 0.02)
)
}
\arguments{
\item{df}{A tibble.}

\item{grp_var}{A categorical variable to group by}

\item{interactive}{If true, plots are passed to plotly and display tooltips.}

\item{n_rows}{How many rows should tbe present in he faceted panel of plots?}

\item{margins}{Margins to be passed to plotly::subplot() if interactive = TRUE.}

\item{cont_var}{A continuous variable.}

\item{idenitifer}{A variable to identify individual points if interactive = TRUE}
}
\description{
Plot boxplots for all continuous variable across zero or one categorical variable.
}
\examples{
plot_box(mtcars, cont_var = mpg, grp_var = cyl, identifier = am)

}
