% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr_target.R
\name{plot_corr_target}
\alias{plot_corr_target}
\title{Plot correlations against target variable}
\usage{
plot_corr_target(
  df,
  target,
  fct_reorder = FALSE,
  fct_rev = FALSE,
  include_lbl = TRUE,
  lbl_precision = 2,
  lbl_position = "outward",
  size = 2,
  line_size = 1,
  vert_size = 1,
  color_pos = palette_light()[[1]],
  color_neg = palette_light()[[2]]
)
}
\arguments{
\item{df}{A tibble}

\item{fct_reorder}{Re-order factors?}

\item{fct_rev}{Reverse factors?}

\item{include_lbl}{Logical scalar - include label in plot?}

\item{lbl_precision}{Numeric scalar indicating placement of label}

\item{size}{Numeric scalar indicating size of plot geoms}

\item{line_size}{Numeric scalar indicating line size for plot}

\item{vert_size}{Numeric scalar indicating size of segment}

\item{color_pos}{Character scalar for color (positive correlations)}

\item{color_neg}{Character scalar for color (negative correlations)}
}
\description{
Plot all correlations against a target variable.
}
\examples{
mtcars \%>\% plot_cor(target = mpg, fct_reorder = T, fct_rev = F, lbl_position = "outward")
}
