% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlogram.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Plot a correlogram}
\usage{
plot_correlogram(df)
}
\arguments{
\item{df}{A tibble}
}
\description{
Plot a correlogram of numeric variables
}
\examples{
plot_correlogram(mtcars)

}
