% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_count}
\alias{plot_count}
\title{Plot Frequencies and Percentages}
\usage{
plot_count(df, interactive = FALSE, ...)
}
\arguments{
\item{df}{A tibble in long form}

\item{interactive}{If true, plots are passed to plotly and display tooltips.}
}
\description{
Make bar plot of frequencies and percentages for one or more categorical variables.
}
\examples{
plot_count(mtcars, cyl)

}
