% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_count_all}
\alias{plot_count_all}
\title{Plot All Frequencies and Percentages}
\usage{
plot_count_all(
  df,
  interactive = FALSE,
  n_rows = 4,
  margins = c(0.059, 0.02, 0.02, 0.02)
)
}
\arguments{
\item{df}{A list of tibbles generated by get_counts_all()}

\item{interactive}{If true, plots are passed to plotly and display tooltips.}

\item{n_rows}{How many rows should tbe present in he faceted panel of plots?}

\item{margins}{Margins to be passed to plotly::subplot() if interactive = TRUE.}
}
\description{
Plot frequencies & pecentages in bar chart for all categorical variables.
}
\examples{
calc_count_all(mtcars) \%>\% plot_count_all()

}
