% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crosscount.R
\name{plot_crosscount}
\alias{plot_crosscount}
\title{Plot Crosscounts}
\usage{
plot_crosscount(df, interactive = FALSE, ...)
}
\arguments{
\item{df}{A tibble, returned from calc_crosscount()}

\item{interactive}{If true, a plotly plot is generated.}
}
\description{
Plot counts and percentages of one categorical variable by another.
}
\examples{
calc_crosscount(mtcars, am, cyl) \%>\% plot_crosscount()

}
