% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crosscount.R
\name{plot_crosscount_all}
\alias{plot_crosscount_all}
\title{Plot Crosscounts for All Categorical Variables Against Target Variable}
\usage{
plot_crosscount_all(df, interactive = FALSE)
}
\arguments{
\item{df}{A tibble}

\item{interactive}{If true, a plotly plot is generated.}
}
\description{
Plot counts and percentages for target categorical variable against all others.
}
\examples{
fht::calc_crosscount_all(mtcars, cyl) \%>\% plot_crosscount_all()

}
