% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Histogram}
\usage{
plot_hist(
  df,
  cont_var,
  bins = 10,
  fill = tidyquant::palette_light()[[3]],
  colour = "white",
  interactive = FALSE
)
}
\arguments{
\item{cont_var}{A continuous variable to plot.}

\item{bins}{A numeric scalar. Number of bins for histograms.}

\item{fill}{A character scalar. Fill colour for histogram bars.}

\item{colour}{A character scalor. Colour for outline of bars}

\item{data}{A tibble.}
}
\description{
Plot a single histogram
}
\examples{
plot_hist(mtcars)

}
