% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist_all}
\alias{plot_hist_all}
\title{Plot Faceted Histograms}
\usage{
plot_hist_all(
  df,
  fct_reorder = FALSE,
  fct_rev = FALSE,
  bins = 10,
  fill = tidyquant::palette_light()[[3]],
  colour = "white",
  ncol = 5,
  scale = "free",
  interactive = FALSE
)
}
\arguments{
\item{fct_reorder}{A logical scalar. Whether factor should retain natural ordering of levels.}

\item{fct_rev}{A logical scalar. Whether factor levels should be reversed}

\item{bins}{A numeric scalar. Number of bins for histograms.}

\item{fill}{A character scalar. Fill colour for histogram bars.}

\item{colour}{A character scalor. Colour for outline of bars}

\item{ncol}{A numeric scalar. Number of columns for faceting panel.}

\item{scale}{A character scalar. Are scales shared across facets for both rows and columns?}

\item{interactive}{If true, plots are passed to plotly and display tooltips.}

\item{data}{A tibble.}
}
\description{
Plot histograms for all continuous variables in dataframe.
}
\examples{
plot_hist_all(mtcars)

}
