% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_scatter.R
\name{plot_scatter_all}
\alias{plot_scatter_all}
\title{Plot all scatterplots}
\usage{
plot_scatter_all(df)
}
\arguments{
\item{df}{A tibble.}
}
\description{
Plot scatterplot for all possible pairs of numeric variables and store in list.
}
\examples{
all_scat<- plot_all_scatter(mtcars) ; all_scat \%>\% purrr::pluck("mpgANDdrat")

}
