% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_test_fun.R
\name{random_test_fun}
\alias{random_test_fun}
\title{Make a random testing fun}
\usage{
random_test_fun(x)
}
\arguments{
\item{x}{A floating point number}
}
\description{
Make random test fun
}
\examples{
random_test_fun(5)

}
