% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_continuous_all.R
\name{summarise_continuous_all}
\alias{summarise_continuous_all}
\title{Summarise Continuous Variables}
\usage{
summarise_continuous_all(df, ...)
}
\arguments{
\item{df}{A tibble.}
}
\description{
Generate a tibble of key statistics for numeric variables.
}
\examples{
summarise_continuous(mtcars)

}
